requirements.txt
altgraph==0.10.2
appnope==0.1.0
backports-abc==0.5
backports.shutil-get-terminal-size==1.0.0
bdist-mpkg==0.5.0
decorator==4.1.2
Django==1.11.11
django-cors-headers==2.1.0
django-crispy-forms==1.7.0
django-filter==1.1.0
django-formtools==2.1
djangorestframework==3.7.7
enum34==1.1.6
future==0.16.0
futures==3.2.0
httplib2==0.9.2
ipython==5.5.0
ipython-genutils==0.2.0
macholib==1.5.1
matplotlib==1.3.1
modulegraph==0.10.4
MySQL-python==1.2.5
mysqlclient==1.3.12
nose==1.3.7
numpy==1.14.2
pathlib2==2.3.0
pexpect==4.3.0
pickleshare==0.7.4
Pillow==5.0.0
prompt-toolkit==1.0.15
ptyprocess==0.5.2
py2app==0.7.3
PyExecJS==1.5.0
Pygments==2.2.0
pyobjc-core==4.1
pyobjc-framework-Cocoa==4.1
pyobjc-framework-CoreData==4.1
pyobjc-framework-Quartz==4.1
pyparsing==2.0.1
python-dateutil==1.5
pytz==2013.7
scandir==1.6
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.11.0
tornado==5.0
traitlets==4.3.2
virtualenv==15.1.0
wcwidth==0.1.7
xattr==0.6.4
zope.interface==4.1.1
